docs: fixed copy/paste text to be correct
The docs have "$" signs for prompts which unfortunately get copied as part of the command (when you hover over the text and press the Copy button). For almost all users, this wouldn't be the desired behavior (i.e. you don't want the $ as part of the command you paste).

Tiny change, hope it helps someone.