docs: fixed yarn instructions and copy/paste to be correct (#116)
* docs: fixed copy/paste text to be correct

The docs have "$" signs for prompts which unfortunately get copied as part of the command (when you hover over the text and press the Copy button). For almost all users, this wouldn't be the desired behavior (i.e. you don't want the $ as part of the command you paste).

Tiny change, hope it helps someone.

* fix: updated documentation for yarn correctness

Two things:
- Updated all references to `yarn` to omit the `$` so that on-hover Copy doesn't include the $
- Updated README to have correct info. IIUC, folks outside Shopify don't have the "dev" command, and should instead use yarn. Also, the instructions for "yarn serve" were incorrect; for local live-serves, you should use `yarn start`.

* docs: twiddle something to retrigger build steps